Renaming Registry
