--- client/src/main/java/cz/vse/java/hangman/controllers/ControllerRegistry.java
@@ -1,69 +1,0 @@
-package cz.vse.java.hangman.controllers;
-
-
-import javafx.stage.Stage;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class ControllerRegistry {
-  private Stage primaryStage;
-
-    private static ControllerRegistry instance;
-
-    private LoginController loginController;
-    private RoomsController roomsController;
-    private CreateRoomController createRoomController;
-    private GameController gameController;
-
-    private static final Logger logger = LoggerFactory.getLogger(ControllerRegistry.class);
-    private ControllerRegistry() {
-
-    }
-    public static ControllerRegistry getInstance() {
-        if (instance == null) {
-            instance = new ControllerRegistry();
-            logger.info("Initializing ControllerRegistry");
-        }
-        return instance;
-    }
-
-    public LoginController getLoginController() {
-        return loginController;
-    }
-
-    public void setLoginController(LoginController loginController) {
-        this.loginController = loginController;
-    }
-
-    public RoomsController getRoomsController() {
-        return roomsController;
-    }
-
-    public void setRoomsController(RoomsController roomsController) {
-        this.roomsController = roomsController;
-    }
-
-    public CreateRoomController getCreateRoomController() {
-        return createRoomController;
-    }
-
-    public void setCreateRoomController(CreateRoomController createRoomController) {
-        this.createRoomController = createRoomController;
-    }
-
-    public GameController getGameController() {
-        return gameController;
-    }
-
-    public void setGameController(GameController gameController) {
-        this.gameController = gameController;
-    }
-
-    public Stage getPrimaryStage() {
-        return primaryStage;
-    }
-
-    public void setPrimaryStage(Stage primaryStage) {
-        this.primaryStage = primaryStage;
-    }
-}
